+++ cluster/resources.tf
@@ -2,13 +2,21 @@ resource "macaddress" "master_net0_mac" {
   count = local.master_count
 }
 
+resource "random_string" "master_serial" {
+  count = local.master_count
+
+  length  = 16
+  special = false
+  upper   = false
+}
+
 module "master_instances" {
   count = local.master_count
 
   source = "github.com/glitchcrab/terraform-module-proxmox-instance?ref=v1.6.0"
 
   pve_instance_name        = "master${count.index}-${local.name_stub}"
-  pve_instance_description = local.master_description
+  pve_instance_description = "${local.master_description} (serial: ${random_string.master_serial[count.index].result})"
   vmid                     = local.vmid_base + count.index + local.vmid_offset
 
   target_node   = lookup(local.host_list[count.index], "name")
@@ -44,13 +52,21 @@ resource "macaddress" "worker_net0_mac" {
   count = local.worker_count
 }
 
+resource "random_string" "worker_serial" {
+  count = local.worker_count
+
+  length  = 16
+  special = false
+  upper   = false
+}
+
 module "worker_instances" {
   count = local.worker_count
 
   source = "github.com/glitchcrab/terraform-module-proxmox-instance?ref=v1.6.0"
 
   pve_instance_name        = "worker${count.index}-${local.name_stub}"
-  pve_instance_description = local.worker_description
+  pve_instance_description = "${local.worker_description} (serial: ${random_string.worker_serial[count.index].result})"
   vmid                     = local.vmid_base + count.index + local.vmid_offset + 5
 
   target_node   = lookup(local.host_list[count.index], "name")