+++ cluster/zz_versions.tf
@@ -10,6 +10,10 @@ terraform {
       source  = "Telmate/proxmox"
       version = "~> 2.9.14"
     }
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 3.4.3"
+    }
     vault = {
       source  = "hashicorp/vault"
       version = "~> 3.11.0"